% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy}
\alias{dummy}
\title{Dummy for a categorial variable}
\usage{
dummy(Y)
}
\arguments{
\item{Y}{: a factor. Data of which to get dummy indicators.}
}
\value{
Dummy-coded matrix-array.
}
\description{
Convert a categorical variable into dummy/indicator variables.
}
\examples{
dummy(iris[,5])
}
