% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pls.R
\name{fit_pls}
\alias{fit_pls}
\title{Partial Least Square Discriminant Analysis Regression}
\usage{
fit_pls(formule, data, n_components = 2)
}
\arguments{
\item{formule}{: object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{: data frame. The data source containing all the variables specified in the formula.}

\item{n_components}{: integer, default=2. Number of components to keep. Should be between 1 and the min(n_samples, n_features, n_targets).}
}
\value{
object of class "PLS" is a list containing the following components for each level of the categorial variable:

formula : The formula given.

X : the explanatory variables (a set of continuous variables).

Y : the target variable (a discrete variable).

Ydum : the dummy-coded matrix linked to the target variable.

n_components : the number of components kept.

Components : a data frame containing the components "th" values for models with 1,...,n_components. Each row corresponds to one individual.

Features_Weights : a data frame containing the weights "wh" values for models with 1,...,n_components. Each row corresponds to one feature.

Coefs_yresiduals_comp : a matrix of coefficients "ch" values for models with 1,...,n_components.

Coefficients_xresiduals_components : a data frame containing the "ph" values for models with 1,...,n_components. Each row corresponds to one feature.

y_residuals : a data frame containing the "yh" residuals values for models with 1,...,n_components. Each row corresponds to one individual.

W_star : a data frame containing the weights "w*h" values for models with 1,...,n_components. Each row corresponds to one feature.

ah : a data frame containing the "ah" coefficients for the regression for models with 1,...,n_components. Each row corresponds to one feature.

u1 : u1,

Coefs_regression : Coefs,

Constant_regression : constant,

all_coefs_regression : Coefs_df,

Matrix_correlation : Cor_var_df,

RSS : RSS,

PRESS : a matrix of PRESS values for models with 1, . . . , n_components.

Q2 : Q2
}
\description{
Fit a PLS-DA Regression.
The PLS-DA is a statistical method used to find the relation between a set of explanatories quantitatives variables and a categorial target variable.
}
\details{
The PLS-DA used consist in a PLS1 Regression.

If the discrete target variable has more than two outcomes, the PLS1 regression is applied on each levels of the categorial variable and using a one versus rest strategy.
}
\examples{
d<-train_test_split(iris,0.7)
train<-d$data_train
objet_pls<-fit_pls(Species~.,train,n_components=4)
objet_pls[[1]]$all_coefs_regression
objet_pls[[2]]$all_coefs_regression
objet_pls[[3]]$all_coefs_regression

}
\references{
TENENHAUSS M., "La régression PLS - Théorie et pratique", Editions Technip

WOLD H. (1985), "Partial Least Squares", in Encyclopedia of Statistical Sciences, vol. 6, Kotz, S & Jonhson, N.L (Eds), John Wiley & Sons, New York, pp.581-591

CHAVENT M. and PATOUILLE B, "Calcul des coefficients de regression et du PRESS en régression PLS1"
}
