% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pls_parall.R
\name{fit_pls_parall}
\alias{fit_pls_parall}
\title{Regression PLS for a data frame}
\usage{
fit_pls_parall(formule, data, n_components = 2, n_cores = 1)
}
\arguments{
\item{formule}{is a formula}

\item{data}{is a data frame}

\item{n_components}{is an integer}

\item{n_cores}{is and integer}
}
\value{
the matrix Wh, Ah, etc
}
\description{
Regression PLS for a data frame
}
\examples{
fit_pls_parall(Species~.,iris,n_components=4,n_cores=1)
}
