% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circle_correlation.R
\name{plot_circle_correlation}
\alias{plot_circle_correlation}
\title{Plotting the circle of correlation for explanatories variables}
\usage{
plot_circle_correlation(objet_pls)
}
\arguments{
\item{objet_pls}{is a class S3}
}
\value{
the plot
}
\description{
Plotting the circle of correlation for explanatories variables
}
\examples{
plot_circle_correlation(fit_pls(Species~.,iris,n_components=4))
}
