% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_explanatories_weights.R
\name{plot_explanatories_weights}
\alias{plot_explanatories_weights}
\title{PLot correlation}
\usage{
plot_explanatories_weights(objet_pls, i = 1)
}
\arguments{
\item{objet_pls}{if a class S3}

\item{i}{is an integer between 1 and the number of outcoms for the target variable}
}
\value{
plot
}
\description{
PLot correlation
}
\examples{
plot_explanatories_weights(fit_pls(Species~.,iris,n_components=4),i=2)
}
