% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_individuals.R
\name{plot_individuals}
\alias{plot_individuals}
\title{plot}
\usage{
plot_individuals(objet_pls, i = 1, j = 1)
}
\arguments{
\item{objet_pls}{is class S3}

\item{i}{integer}

\item{j}{integer}
}
\value{
plot
}
\description{
plot
}
\examples{
plot_individuals(fit_pls(Species~.,iris,n_components=4),i=2,j=1)
}
