% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Q2.R
\name{plot_Q2}
\alias{plot_Q2}
\title{Q2 plot}
\usage{
plot_Q2(objet_pls)
}
\arguments{
\item{objet_pls}{: object of class S3 inheriting from "PLS".}
}
\value{
the graph of the Q2 values for all the strategies one versus all.

A threshold of 0.0975 is drawn in red on the graph.
}
\description{
Plots the Q2 values for all the outcomes and the number of components selected previously.
}
\examples{
plot_Q2(fit_pls(Species~.,iris,n_components=4))
}
