% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pls.R
\name{predict_pls}
\alias{predict_pls}
\title{predictions}
\usage{
predict_pls(objet_pls, newdata, type = 1)
}
\arguments{
\item{objet_pls}{is class S3}

\item{newdata}{is a data frame}

\item{type}{1 pour ou 2}
}
\value{
predictions
}
\description{
predictions
}
\examples{
d<-train_test_split(iris,0.8)
train<-d$data_train
test<-d$data_test
res<-fit_pls(Species~.,train,3)
predict_pls(res,test,2)
}
