% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PLS.R
\name{print.PLS}
\alias{print.PLS}
\title{Printing PLS-DA Fits}
\usage{
\method{print}{PLS}(x, ...)
}
\arguments{
\item{x}{: object of class S3 inheriting from "PLS".}

\item{...}{: other parameters.}
}
\value{
print returns a list containing the following components for each level of the categorical variable:

data frame of RSS values for models with 1, . . . , n_components.

data frame of PRESS values for models with 1, . . . , n_components.

data frame of Wh values for models with 1, . . . , n_components.

data frame of Ch values for models with 1, . . . , n_components.

data frame of Ph values for models with 1, . . . , n_components.

data frame of Wh* values for models with 1, . . . , n_components.

data frame of Ah values for models with 1, . . . , n_components.
}
\description{
Printing PLS-DA Fits
}
\examples{
d<-train_test_splits(iris,0.7)
train<-d$data_train
objet<-fit_pls(Species~.,train,n_components=4)
print(objet)
print(objet)$RSS

m<-train_test_splits(iris,0.8)
train1<-m$data_train
obj<-fit_pls(Species~Petal.Length+Sepal.Width,train1,n_components=2)
print(obj)
print(objet)$Ah
}
