% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_splits.R
\name{train_test_splits}
\alias{train_test_splits}
\title{Train and test split data}
\usage{
train_test_splits(data, train_size = 0.6)
}
\arguments{
\item{data}{: data frame.}

\item{train_size}{: float, default=0.6.
Should be a float between 0.0 and 1.0 and represent the proportion of the data frame to include in the test split.}
}
\value{
List of data frame containing train-test split of inputs ($data_train and $data_test).
}
\description{
Split a data frame into random train and test subsets.
}
\examples{
train_test_splits(mtcars)
d<-train_test_splits(iris,0.8)
train<-d$data_train
test<-d$data_test
train
test
}
