% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xscale.R
\name{xscale}
\alias{xscale}
\title{Scale data}
\usage{
xscale(X, reduce = TRUE)
}
\arguments{
\item{X}{: matrix or data frame.}

\item{reduce}{: bool, default = TRUE.
If True, scale the data to unit standard deviation.}
}
\value{
a matrix-array of the transformed data.
}
\description{
Standardize each column of a data set.
}
\examples{
xscale(iris[,1:4])
xscale(iris[,1:4],reduce=FALSE)
}
