% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ydummy.R
\name{ydummy}
\alias{ydummy}
\title{Dummy for a categorial variable}
\usage{
ydummy(Y)
}
\arguments{
\item{Y}{: a factor. Data of which to get dummy indicators.}
}
\value{
Dummy-coded matrix-array.
}
\description{
Convert a categorical variable into dummy/indicator variables.
}
\examples{
ydummy(iris[,5])
}
