% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circle_correlation.R
\name{plot_circle_correlation}
\alias{plot_circle_correlation}
\title{Circle of correlation}
\usage{
plot_circle_correlation(objet_pls)
}
\arguments{
\item{objet_pls}{: object of class S3 inheriting from "PLS".}
}
\value{
the graph of the circle of correlation for all the features.
}
\description{
Plots the circle of correlation
}
\examples{
d<-train_test_splits(iris,0.7)
train<-d$data_train
objet<-fit_pls(Species~.,train,n_components=4)
plot_circle_correlation(objet)
}
