% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_individuals_components.R
\name{plot_individuals_components}
\alias{plot_individuals_components}
\title{Maps individuals components}
\usage{
plot_individuals_components(objet_pls, i = 1)
}
\arguments{
\item{objet_pls}{is a class S3}

\item{i}{is an integer}
}
\value{
plot
}
\description{
Maps individuals components
}
\examples{
plot_individuals_components(fit_pls(Species~.,iris,n_components=4),i=2)
}
