% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Q2.R
\name{plot_Q2}
\alias{plot_Q2}
\title{PLot Q2}
\usage{
plot_Q2(objet_pls)
}
\arguments{
\item{objet_pls}{is a class S3}
}
\value{
a plot
}
\description{
PLot Q2
}
\examples{
plot_Q2(fit_pls(Species~.,iris,n_components=4))
}
