% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pls.R
\name{predict_pls}
\alias{predict_pls}
\title{Predict}
\usage{
predict_pls(objet_pls, newdata, type = 1)
}
\arguments{
\item{objet_pls}{: object of class S3 inheriting from "PLS".}

\item{newdata}{: data frame. The data test source containing all the variables specified in the formula. 'Na's are not allowed.}

\item{type}{: integer equal to 1 or 2. 1 to predict the class and 2 in order to add the estimation of the class membership probabilities.}
}
\value{
a list of predicted class and the estimation of the class membership probabilities if type=2.
}
\description{
Predicts the class of an individuals given the results of the PLS-DA model fitting function.
}
\note{
The softmax function is used for the estimation of the class membership probabilities if type=2.

The argmax is used to predict the class.
}
\examples{
d<-train_test_split(iris,0.8)
train<-d$data_train
test<-d$data_test
res<-fit_pls(Species~.,train,3)
predict_pls(res,test,2)

}
