% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pls.R
\name{print.PLS}
\alias{print.PLS}
\title{Print PLS}
\usage{
\method{print}{PLS}(x, ...)
}
\arguments{
\item{x}{is a class S3}

\item{...}{other parameters}
}
\value{
print all the matrix Ah, ...
}
\description{
Print PLS
}
\examples{
print(fit_pls(Species~.,iris,n_components=4))
}
