% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pls.R
\name{summary.PLS}
\alias{summary.PLS}
\title{Summary}
\usage{
\method{summary}{PLS}(object, ...)
}
\arguments{
\item{object}{S3 class PLS}

\item{...}{other parameters}
}
\value{
summary
}
\description{
Summary
}
\examples{
summary(fit_pls(Species~.,iris,n_components=2))
}
