% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_predict.R
\name{summary_predict}
\alias{summary_predict}
\title{summary pls predict}
\usage{
summary_predict(pls_predict)
}
\arguments{
\item{pls_predict}{is an object}
}
\value{
confusion matrix
}
\description{
summary pls predict
}
\examples{
d<-train_test_split(iris,0.8)
train<-d$data_train
test<-d$data_test
res<-fit_pls(Species~.,train,3)
pred<-predict_pls(res,test,2)
summary_predict(pred)
}
